+++ src/components/menu/__internal__/submenu/submenu.test.tsx
@@ -4,7 +4,7 @@ import userEvent from "@testing-library/user-event";
 
 import FixedNavigationBarContext from "components/navigation-bar/fixed-navigation-bar.context";
 import { Menu, MenuItem, MenuSegmentTitle } from "../..";
-import MenuContext, { MenuContextProps } from "../../menu.context";
+import MenuContext, { MenuContextProps } from "../menu.context";
 import Submenu from "./submenu.component";
 import ScrollableBlock from "../../scrollable-block";
