+++ src/components/menu/menu-divider/menu-divider.component.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import StyledDivider from "./menu-divider.style";
-import MenuContext from "../menu.context";
+import MenuContext from "../__internal__/menu.context";
 import { StyledMenuItem } from "../menu.style";
 import tagComponent, {
   TagProps,
