+++ src/components/menu/menu-divider/menu-divider.style.ts
@@ -1,7 +1,7 @@
 import styled, { css } from "styled-components";
 import menuConfigVariants from "../menu.config";
 import { MenuDividerProps } from "./menu-divider.component";
-import { MenuType } from "../menu.context";
+import { MenuType } from "../__internal__/menu.context";
 
 const StyledDivider = styled.div<MenuDividerProps & { menuType: MenuType }>`
   cursor: default;
