+++ src/components/vertical-divider/vertical-divider.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import { testStyledSystemSpacing } from "../../__spec_helper__/__internal__/test-utils";
 import VerticalDivider from ".";
-import { MenuContext } from "../menu/menu.component";
+import MenuContext from "../menu/__internal__/menu.context";
 
 testStyledSystemSpacing((props) => <VerticalDivider {...props} />, {
   p: 3,
